requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.7.1
apturl==0.5.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.6.3
Brlapi==0.8.3
bs4==0.0.1
cachetools==5.0.0
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cssselect==1.2.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.4.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
elastic-transport==8.4.0
elasticsearch==7.11.0
exceptiongroup==1.1.3
Faker==19.3.1
fastapi==0.101.1
fasteners==0.14.1
filelock==3.12.2
Flask==2.1.0
flask-restplus==0.13.0
flask-restx==0.5.1
frozenlist==1.4.0
future==0.18.2
gevent==23.9.0.post1
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
googletrans==4.0.0rc1
greenlet==3.0.0
gyp==0.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kafka-python==2.0.2
keyring==23.5.0
langdetect==1.0.9
langid==1.1.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libretranslatepy==2.1.1
lockfile==0.12.2
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
numpy==1.25.2
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
outcome==1.2.0
pandas==2.0.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.10.0
protobuf==3.12.4
ptyprocess==0.7.0
py3langid==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycups==2.0.1
pydantic==2.2.1
pydantic_core==2.6.1
PyDrive==1.3.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.4.1
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyquery==2.0.0
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
pytz==2023.3.post1
pyxdg==0.27
PyYAML==6.0.1
referencing==0.30.2
reportlab==3.6.8
requests==2.30.0
rfc3986==1.5.0
rpds-py==0.10.6
rsa==4.8
SecretStorage==3.3.1
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.31
starlette==0.27.0
systemd-python==234
texttable==1.6.4
translate==3.6.1
trio==0.22.2
trio-websocket==0.10.3
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.18
usb-creator==0.3.7
uvicorn==0.23.2
vboxapi==1.0
virtualenv==20.24.2
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==2.1.2
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
zope.event==5.0
zope.interface==6.1
